Add new blueprints for the logs routes and the routes that will be scheduled. Refactor routes to a blueprints folder. Remove unused dependencies from the package.json and component files.
