+++ frontend/src/app/modules/logs/logs.component.html
@@ -37,6 +37,6 @@
             </tr>
         </div>
     </table>
-    <mat-paginator [pageSizeOptions]="[5, 10, 20]" showFirstLastButtons></mat-paginator>
+    <mat-paginator [pageSizeOptions]="[20, 100, 150]" showFirstLastButtons></mat-paginator>
     <mat-spinner class="spinner" *ngIf="showSpinner" style="position:fixed"></mat-spinner>
 </div>
